WIP to auto-name 'mons to route name
Does not function perfectly yet, keeps naming the pokemon gibberish.

If multiple pokemon are caught on the same route, they will be named the same, which shows that something is working. Perhaps we need to parse throuh a route name list to get the correct data before saving it to the pokemon.